feat: add error handling messages
