+++ graphiql-app/src/app/[locale]/globals.css
@@ -28,7 +28,6 @@ main {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  gap: 10vh;
   max-width: 100vw;
 }
