+++ graphiql-app/src/app/[locale]/graphql/error.tsx
@@ -1,18 +1,8 @@
 'use client';
 
-import Link from 'next/link';
-import { Container, Typography } from '@mui/material';
+import ErrorPage from '@/components/ErrorPage/ErrorPage';
+import ErrorProps from '@/types/types';
 
-export default function Error() {
-  return (
-    <Container sx={{ textAlign: 'center' }} maxWidth="md" component="main">
-      <Typography component="h1" variant="h3" gutterBottom>
-        ERROR PAGE
-      </Typography>
-
-      <Link className="Link" href="welcome">
-        Go to welcome
-      </Link>
-    </Container>
-  );
+export default function Error({ error, reset }: ErrorProps) {
+  return <ErrorPage error={error} reset={reset} />;
 }
